pyproject.toml
[tool.pdm]

[project]
name = "nonebot_plugin_acm_reminder"
version = "0.1.6"
description = "A subscribe CodeForces/NowCoder contest info plugin for nonebot2"
authors = [
    {name = "BigOrangeQWQ", email = "2284086963@qq.com"},
]
dependencies = [
    "BeautifulSoup4>=4.11.2",
    "nonebot2>=2.0.0b5",
    "nonebot-adapter-onebot>=2.1.3",
    "pydantic>=1.9.2",
    "httpx>=0.23.3",
    "nonebot-plugin-apscheduler>=0.2.0",
    "nonebot-plugin-htmlrender>=0.2.0.3",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=1.0"]
build-backend = "pdm.pep517.api"
